Added Hazard Markers

+++ app/src/main/java/com/hazardiqplus/data/SaveHazardRequest.kt
@@ -1,7 +1,7 @@
 package com.hazardiqplus.data
 
 data class SaveHazardRequest(
-    val rad: Int,
+    val rad: Double,
     val lat: Double,
     val lng: Double,
     val hazard: String
